% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brass_mort.R
\name{brass_mort}
\alias{brass_mort}
\title{Brass's Relational Model of Mortility}
\usage{
brass_mort(model = NULL, x = seq(from = 0, to = 100, by = 1), alpha = 0,
  beta = 1, l0 = 1e+05)
}
\arguments{
\item{model}{Vector of a `model` age specific person years lived schedule $L_x$. Will be used to derive the standardised transformed age schedule, \eqn{\Y_s(x)} above.}

\item{x}{Vector for the sequence of ages.}

\item{alpha}{Numeric value for intercept adjustment to the standardised transformed age schedule}

\item{beta}{Numeric value for slope adjustment to the standardised transformed age schedule}
}
\value{
Returns vector of $L_x$ from the relational model schedule. The function is primarily intended for creating synthetic survivorships for projection models.

The \code{alpha} and \code{beta} parameters relate a modified schedule to the schedule provided by the \code{model} argument.
}
\description{
Brass' relational model of mortility makes use of a standard model mortility schedule, based on $L_x^s$ that is linearised using
\deqn{Y_s(x) = \frac{1}{2} \ln \left( \frac{a l_0 - L_x^s}{L_x^s} \right) }
where \eqn{l_0} is the population of the cohort at age zero.
}
\details{
The modified transformed schedule is formed by adjusting the demographer's logit above using
\deqn{Y(x) = \alpha + \beta Y_s(x)}
where \eqn{\beta} changes the slope and \eqn{\alpha} the intercept of the linearised $L_x$

The modified transformed schedule is converted back into $L_x$ by applying the anti-function used to linearise the standard schedule.
}
\examples{
#single year
df0 <- subset(austria, Year == 2014)
f0 <- df0$Lx_f
f1 <- brass_mort(model = f0, x = df0$Age, alpha = 0.8, beta = 1)
plot(f0, type = "l")
lines(f1, col = "red")
e0 <- sum(Lx)/l0
e0
}
\author{
Guy J. Abel
}
