% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage.R
\name{gage}
\alias{gage}
\title{Gage`s Extenstion to Brass` Polynomial Fertility Schedule}
\usage{
gage(tfr = NULL, x = seq(from = 0, to = 100, by = 1), start_fertage = 15,
  width_fertage = 35, c = 12 * w^-4 * tfr, scaled = TRUE)
}
\arguments{
\item{tfr}{Numeric value for total fertitliy rate of the returned age schedule.}

\item{x}{Vector for the sequence of ages.}

\item{start_fertage}{Numeric value for the start of the fertility age range. Equivalent to \eqn{s} in the equation above.}

\item{width_fertage}{Numeric value for the width of the fertility age range. Equivalent to \eqn{w} in the equation above.}

\item{c}{Numeric value for a constant to control the overall fertility level. By default calcuated to approxiamate value in \code{tfr} argument.}

\item{scaled}{Boolean value to return age specific fertility rates that scale to sum to \code{tfr} or not. Default \code{TRUE}.}
}
\value{
Returns the f(x) values from the Gage extenstion to Brass` schedule of age specific fertility. The age range for the calculation can take any sequence of positive numbers, such as ages in single or 5-year intervals. The function is primarily intended for use in decomposing a total fertility rate into an age-specific values.

The parameters relate to the start and width of the fertility age range (\code{start_fertage}, \code{ width_fertage}) and the oveall level (\code{tfr}, \code{c}). By default \code{s} is set to 15 years and \code{w} to 33 as fixed in Brass' original schedule.

The adapted version in this function, to allow scaling to a known TFR, was based on IUSSP Demographic Models online section (see \url{http://papp.iussp.org/sessions/papp103_s03/PAPP103_s03_040_040.html} for more details).
}
\description{
Provides a scaled Gage fertility schedule,
\deqn{ f(x) = c (x - s) (s + w - x) ^ 2}
for a given set of parameter values and sequence of ages.
}
\examples{
#single year
f0 <- gage(tfr = 1.8)
plot(f0, type = "l")

#five year
f1 <- gage(tfr = 5.8, x = seq(from = 0, to = 100, by = 5))
plot(f1, type = "l")
sum(f1)
}
\author{
Guy J. Abel
}
\references{
Gage, T. B. (2001). `Age-specific fecundity of mammalian populations: A test of three mathematical models`. \emph{Zoo Biology} 20(6):487-499.

Brass, W. (1975). \emph{Methods for Estimating Fertility and Mortality from Limited and Defective Data.} Chapel Hill NC: Carolina Population Centre.
}
\seealso{
\code{\link{.}}
}

