% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_fert.R
\name{gamma_fert}
\alias{gamma_fert}
\title{Gamma Fertility Model Schedule.}
\usage{
gamma_fert(tfr = NULL, x = seq(from = 0, to = 100, by = 1),
  mean_cb = NULL, shape = NULL, start_fertage = 15, width_fertage = 35,
  ...)
}
\arguments{
\item{tfr}{Numeric value for total fertitliy rate of the returned age schedule.}

\item{x}{Vector for the sequence of ages.}

\item{mean_cb}{Numeric value for the mean age of mothers at child birth. Equivalent to \eqn{k\theta}.}

\item{shape}{Numeric poisitive value for schedule shape. Used direcly in \code{\link{dgamma}}.}

\item{start_fertage}{Numeric value for the start of the fertility age range. Equivalent to \eqn{\alpha} in the equation above.}

\item{width_fertage}{Numeric value for the width of the fertility age range. Equivalent to \eqn{\delta} in the equation above.}

\item{...}{Additional arguments passed to \code{\link{dgamma}}}
}
\value{
Returns the f(x) values from a gamma schedule of age specific fertility. The age range for the calculation can take any sequence of positive numbers, such as ages in single or 5-year intervals. The function is primarily intended for use in decomposing a total fertility rate into an age-specific values.

The parameters relate the mean age of child bearing (\code{mean_cb}) and the shape of the schedule (\code{shape}). The shape and mean age parameters are used to determine the scale parameter of the gamma distribution.

The arguments for the start and width of the fertility age range (\code{start_fertage}, \code{width_fertage}) are used to select where the gamma distribution is applied over the range of ages given in \code{x}.
}
\description{
Provides a scaled gamma fertility schedule,
\deqn{ f(x) = \frac{1}{\Gamma(k)\theta^k} x^{(k-1)}e^{-\frac{x}{\theta}} }
for a given set of parameter values and sequence of ages.
}
\examples{
#single year
f0 <- gamma_fert(tfr = 4, mean_cb = 25, shape = 5)
plot(f0, type = "l")

#five year
f1 <- gamma_fert(tfr = 5.8, mean_cb = 28, x = seq(from = 0, to = 100, by = 5), shape = 3)
plot(f1, type = "l")
sum(f1) * 5
}
\seealso{
\code{\link{gage}} \code{\link{romainuk}} \code{\link{hadwiger}}
}
\author{
Guy J. Abel
}
