% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadwiger.R
\name{hadwiger}
\alias{hadwiger}
\title{Hadwiger Fertility Model Schedule.}
\usage{
hadwiger(tfr = NULL, x = seq(from = 0, to = 100, by = 1), H = NULL,
  T = NULL, d = NULL, start_fertage = 15, width_fertage = 35,
  scaled = TRUE)
}
\arguments{
\item{tfr}{Numeric value for total fertitliy rate of the returned age schedule.}

\item{x}{Vector for the sequence of ages.}

\item{H}{Numeric value for the \eqn{H} parameter in the equation above.}

\item{T}{Numeric value for the \eqn{T} parameter in the equation above.}

\item{d}{Numeric value for the \eqn{d} parameter in the equation above.}

\item{start_fertage}{Numeric value for the start of the fertility age range.}

\item{width_fertage}{Numeric value for the width of the fertility age range.}

\item{scaled}{Boolean value to return age specific fertility rates that scale to sum to \code{tfr} or not. Default \code{TRUE}.}
}
\value{
Returns the f(x) values from the Hadwiger schedule of age specific fertility, later refined by Gilje and Yntema. The age range for the calculation can take any sequence of positive numbers, such as ages in single or 5-year intervals. The function is primarily intended for use in decomposing a total fertility rate into an age-specific values.

The parameters relate to the shape of the schedule (\code{H}, \code{T} and \code{d}). The shape parameters have no demographic interpretion.

The arguments for the start and width of the fertility age range (\code{start_fertage}, \code{width_fertage}) are used to select where the distribution is applied over the range of ages given in \code{x}.

The adapted version in this function, to allow scaling to a known TFR, was based on IUSSP Demographic Models online section (see \url{http://papp.iussp.org/sessions/papp103_s03/PAPP103_s03_040_090.html} for more details).
}
\description{
Provides a scaled Hadwiger fertility schedule,
\deqn{f(x) = \frac{H}{T \sqrt{\pi}}  \left( \frac{T}{x - d} \right) ^ \frac{3}{2}
            \exp \left\{-H^2 \left(\frac{T}{x-d} \frac{x-d}{T} -2\right) \right\} }
for a given set of parameter values and sequence of ages.
}
\examples{
#single year
f0 <- hadwiger(tfr = 1.8, H = 4, T = 28, d = 2)
plot(f0, type = "l")

#five year
f1 <- hadwiger(tfr = 5.8, x = seq(from = 0, to = 100, by = 5), H = 3, T = 26, d = 2, width_fertage = 40)
plot(f1, type = "l")
sum(f1) * 5
}
\references{
Gilje E. and Yntema L. (1970). `The shifted Hadwiger fertility function`. Working Paper No IO70/14. \emph{Oslo: Central Bureau of Statistics of Norway.}

Hadwiger H. (1940). Eine analytische Reprodutionsfunktion fur biologische Gesamtheiten. \emph{Skandinavisk Aktuarietidskrift} 23(101-113)
}
\seealso{
\code{\link{.}}
}
\author{
Guy J. Abel
}
