% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romaniuk.R
\name{romaniuk}
\alias{romaniuk}
\title{Romaniuk’s Three-Parameter Model.}
\usage{
romaniuk(tfr = 2.1, A = NULL, M = NULL, alpha = NULL, delta = NULL,
  ag = 1)
}
\arguments{
\item{tfr}{Numeric value for total fertitliy rate of the returned age
age schedule}

\item{A}{Numeric value for average age parameter.}

\item{M}{Numeric value for modal age parameter.}

\item{alpha}{Numeric value for start of fertile age range.}

\item{ag}{Numeric value for age groupings. \code{1} year age groups used by default.}

\item{beta}{Numeric value for duration of fertile age range.}
}
\value{
A vector for the age specific fertility rates generated from the
  Romaniuk age schedule. See
  \url{http://papp.iussp.org/sessions/papp103_s03/PAPP103_s03_040_070.html} for more details.
}
\description{
Returns a fertility age schedule using the paramaterisation of
Romainuk (1973) scaled to have a total fertitliy rate
(T) matching the \code{tfr} argument;
\deqn{ f(x) = T \left\( 1 + \frac{x}{M-\alpha}            \right\)^\left\{ \frac{(M-\alpha)         [\beta-2(A-\alpha)]}{\beta(A-M)} \right\}
               \left\( 1 - \frac{x}{\delta - M + \alpha} \right\)^\left\{ \frac{(\delta - M+\alpha)[\beta-2(A-\alpha)]}{\beta(A-M)} \right\}  }
}
\examples{
#single year
f0 <- romaniuk(tfr = 1.8, A = 28.2, M = 26, alpha = 15, delta = 33, a = 1)
plot(f0)

#five year
f1 <- romaniuk(tfr = 5.8, A = 27, M = 23, alpha = 15, delta = 35, a = 5)
plot(f1)
}
\author{
Guy J. Abel
}
\references{
Romaniuk, Anatole (1973). "A three parameter mode for birth projections",
 \emph{Population Studies} 27(3):467-478. doi 10.2307/2173766
}
\seealso{
\code{\link{.}}
}

