% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brass_mort.R
\name{brass_mort}
\alias{brass_mort}
\title{Brass's Relational Model of Mortility}
\usage{
brass_mort(model = NULL, x = seq(from = 0, to = 100, by = 1), alpha = 0,
  beta = 1, l0 = 1e+05)
}
\arguments{
\item{model}{Vector of a `model` age specific mortility rates. Will be used to derive the standardised transformed age schedule, \eqn{\Y_s(x)} above.}

\item{x}{Vector for the sequence of ages.}

\item{alpha}{Numeric value for intercept adjustment to the standardised transformed age schedule}

\item{beta}{Numeric value for slope adjustment to the standardised transformed age schedule}
}
\value{
Returns a life table with values from a relational model schedule of age specific mortility. The age range for the calculation can take any sequence of positive numbers, such as ages in single or 5-year intervals. The function is primarily intended for creating synthetic mortility rate into an age-specific values.

The \code{alpha} and \code{beta} parameters relate a modified schedule to the schedule provided by the \code{model} argument.
}
\description{
Brass' relational Gompertz model of mortility makes use of a model mortility schedule that is linearised using
\deqn{Y_s(x) = \frac{1}{2} \ln \left( \frac{l^{*}_x}{1-l^{*}_x} \right) }
where \eqn{l^{*}_x} is the proportions surviving at exact age x, i.e. the $l_x$ from the life table scaled to sum to a radix of 1.
}
\details{
The modified transformed schedule is formed by adjusting the demographer's logit above using
\deqn{Y(x) = \alpha + \beta Y_s(x)}
where \eqn{\beta} changes the slope and \eqn{\alpha} the intercept of the linearised $l_x$

The modified transformed schedule is converted back into $l_x$ by applying the anti-function used to linearise the standard schedule and multiplying by original radix of $l_x$
}
\examples{
#single year
df0 <- subset(austria, Year == 2014)
df0$lx_f <- df0$Lx_f + df0$Dx_f/2
f0 <- df0$lx_f

f1 <- brass_mort(model = f0, x = seq(from = 0, to = 110, by = 1), alpha = 0, beta = 1)
plot(f0, type = "l")
lines(f1$lx, col = "red")
f1$ex[1]
}
\author{
Guy J. Abel
}
