% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fert.R
\name{model_fert}
\alias{model_fert}
\title{Model Fertility Model Schedule.}
\usage{
model_fert(tfr = NULL, x = seq(from = 0, to = 100, by = 1), model = NULL,
  start_fertage = 15, width_fertage = 35)
}
\arguments{
\item{tfr}{Numeric value for total fertitliy rate of the returned age schedule.}

\item{x}{Vector for the sequence of ages.}

\item{model}{Vector of a `model` age specific fertility rates.}

\item{start_fertage}{Numeric value for the start of the fertility age range.}

\item{width_fertage}{Numeric value for the width of the fertility age range.}
}
\value{
Returns the f(x) values from a model schedule of age specific fertility. The age range for the calculation can take any sequence of positive numbers, such as ages in single or 5-year intervals. The function is primarily intended for use in decomposing a total fertility rate into an age-specific values.

The arguments for the start and width of the fertility age range (\code{start_fertage}, \code{width_fertage}) are used to select where the model distribution is applied over the range of ages given in \code{x}. Ensure that these match those of the model age schedule.
}
\description{
Model Fertility Model Schedule.
}
\examples{
#single year
f0 <- gamma_fert(tfr = 4, mean_cb = 25, shape = 5)
plot(f0, type = "l")

#five year
f1 <- gamma_fert(tfr = 5.8, x = seq(from = 0, to = 100, by = 5), shape = 3)
plot(f1, type = "l")
sum(f1)
}
\author{
Guy J. Abel
}
\seealso{
\code{\link{gage}} \code{\link{romainuk}} \code{\link{hadwiger}}
}

